Adds suggested changes, adds two examples.

+++ code/modules/admin/view_variables/topic.dm
@@ -529,20 +529,22 @@
 	else if(href_list["addaura"])
 		if(!check_rights(R_DEBUG|R_ADMIN|R_FUN))	return
 		var/mob/living/L = locate(href_list["addaura"])
+		if(!istype(L))
+			return
 		var/choice = input("Please choose an aura to add", "Auras", null) as null|anything in typesof(/obj/aura)
 		if(!choice || !L)
 			return
 		var/obj/o = new choice(L)
-		log_admin("[key_name(usr)] added [o] to \the [L].")
-		message_admins("<span class='notice'>[key_name(usr)] added [o] to \the [L].</span>")
+		log_and_message_admins("added \the [o] to \the [L]")
 	else if(href_list["removeaura"])
 		if(!check_rights(R_DEBUG|R_ADMIN|R_FUN))	return
 		var/mob/living/L = locate(href_list["removeaura"])
+		if(!istype(L))
+			return
 		var/choice = input("Please choose an aura to remove", "Auras", null) as null|anything in L.auras
 		if(!choice || !L)
 			return
-		log_admin("[key_name(usr)] removed [choice] from \the [L].")
-		message_admins("<span class='notice'>[key_name(usr)] removed [choice] from \the [L].</span>")
+		log_and_message_admins("removed \the [choice] to \the [L]")
 		qdel(choice)
 	if(href_list["datumrefresh"])
 		var/datum/DAT = locate(href_list["datumrefresh"])
